created a class for server and made it so messages can now be sent between devices with usernames hopefully, however need to press enter lots of time and still gets disconnect errors.
